Optionally disable pulling in the tokio runtime
Before this patch, hyper-rustls required pulling in the tokio runtime
due to it's use of `hyper::client::HttpConnector`. Unfortunately
Fuchsia does not support tokio, which prevents our use of this
library. This PR allows us to optionally disable pulling in tokio,
and instead swap in our own version of the hyper `HttpConnector`.