don't depend on hyper/tcp feature by default
On fuchsia, we don't support the mio and tokio runtime yet, and so in order
to use hyper-rustls, we need to compile it with `default-features = false`.
Unfortunately, in https://github.com/ctz/hyper-rustls/pull/96, @CryZe added
a hard requirement for the hyper "tcp" feature, which pulls in the tokio
runtime.

To restore our ability to use hyper-rustls, this patch removes the
"hyper/tcp" hard feature requirement.