Fix bad assert on alpn_protocols and add tests
An assert was added on WantsProtocols1.wrap_connector which was the
wrong place, as we update the alpn_protocols field ourselves before
calling wrap_connector.  Asserting on with_tls_config is sufficient,
this is the only place ConnectorBuilder(WantsSchemes) is constructed
and the builder will always pass here.

Add tests for alpn_protocols and the expected panic when trying to force
it from outside.