chore: avoid fixed sleep in integration tests.
Prior to this commit the `tests/tests.rs` integration tests had an
unconditional 1 second sleep waiting for a server to become available.
This is slightly problematic:

* Waiting a full second if the server becomes available sooner is
  unnecessarily slow.
* Conversely, we can't be sure the server is available after the 1s
  sleep.

This commit replaces the sleeps with a loop of up to 10 iterations that
tries to connect to the server port. If any iteration succeeds we return
immediately to avoid unnecessary waiting. If an iteration fails, we wait
longer, increasing the duration each time we wait (up to a maximum of 10
iteration).