fix: kill server before asserting on output.
Prior to this commit the `server()` test would call `assert_eq!` to
assert on client output matching expected _before_ calling `srv.kill()`
on the `Child` spawned by `server_command()`.

The upstream rustdocs for `std.process.Child` mention:
> There is no implementation of Drop for child processes, so if you do
> not ensure the Child has exited then it will continue to run, even
> after the Child handle to the child process has gone out of scope.

The net result is that if the assertion in `server()` fails, we leave
a running server instance behind. This manifests as a port conflict when
re-running the test without first killing the orphan.

This commit moves the `srv.kill()` call to before the `assert_eq`. This
avoids leaking the server process if the assertion fails.