CI: replace ubuntu 18.04 w/ 20.04.
The Build+Test matrix job of the GitHub actions workflow was using
Ubuntu 18.04, which is deprecated.

This commit replaces it with Ubuntu 20.04, the oldest supported Ubuntu
release available in GitHub actions at the time of writing.

The other workflow task are using `ubuntu-latest` and do not need
updating at this time.