Cargo: update to rustls 0.22
This commit updates to rustls 0.22, taking the following associated
updates:
* rustls 0.22.0-alpha-6 -> 0.22
* pki-types 0.2 -> 1
* tokio-rustls 0.25.0-alpha.4 -> 0.25
* rustls-native-certs 0.7.0-alpha.3 -> 0.7
* webpki-roots 0.26.0-alpha.2 -> 0.26
* rustls-pemfile 2.0.0-alpha.2 -> 2

Breaking API changes are addressed as required. Notably, the builder fns
that accept a custom provider and use the safe defaults are now fallible
to account for a possible error if the provider's configuration is not
compatible with the default safe protocol versions.